move AccountActivity and related classes to a new package
