+++ app/src/main/java/com/keylesspalace/tusky/di/FragmentBuildersModule.kt
@@ -29,7 +29,7 @@ import com.keylesspalace.tusky.components.search.fragments.SearchHashtagsFragmen
 import com.keylesspalace.tusky.components.search.fragments.SearchStatusesFragment
 import com.keylesspalace.tusky.components.timeline.TimelineFragment
 import com.keylesspalace.tusky.fragment.AccountListFragment
-import com.keylesspalace.tusky.fragment.AccountMediaFragment
+import com.keylesspalace.tusky.components.account.media.AccountMediaFragment
 import com.keylesspalace.tusky.fragment.NotificationsFragment
 import com.keylesspalace.tusky.fragment.ViewThreadFragment
 import dagger.Module
