+++ app/src/main/java/com/keylesspalace/tusky/di/ViewModelFactory.kt
@@ -12,7 +12,7 @@ import com.keylesspalace.tusky.components.report.ReportViewModel
 import com.keylesspalace.tusky.components.scheduled.ScheduledTootViewModel
 import com.keylesspalace.tusky.components.search.SearchViewModel
 import com.keylesspalace.tusky.components.timeline.TimelineViewModel
-import com.keylesspalace.tusky.viewmodel.AccountViewModel
+import com.keylesspalace.tusky.components.account.AccountViewModel
 import com.keylesspalace.tusky.viewmodel.AccountsInListViewModel
 import com.keylesspalace.tusky.viewmodel.EditProfileViewModel
 import com.keylesspalace.tusky.viewmodel.ListsViewModel
